requirements.txt
flask==3.0.0
flask-restx==1.3.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
mlflow==2.9.2
evidently==0.7.17
streamlit==1.29.0
requests==2.31.0
werkzeug==3.0.1
pytest==7.4.3
apache-airflow==2.8.0
